requirements.txt
anthropic==0.40.0
appdirs==1.4.4
atomicwrites==1.4.0
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
filetype==1.2.0
groq==0.13.0
http-exceptions==0.2.10
httpx-sse==0.4.0
jiter==0.8.0
jsonpointer==2.1
langchain-anthropic==0.3.0
langchain-core==0.3.21
langchain-groq==0.2.1
langchain-openai==0.2.10
langgraph==0.2.53
langgraph-checkpoint==2.0.7
langgraph-sdk==0.1.40
langsmith==0.1.147
llama-cloud==0.1.5
llama-index==0.12.2
llama-index-agent-openai==0.4.0
llama-index-cli==0.4.0
llama-index-core==0.12.2
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.3.2
llama-index-multi-modal-llms-openai==0.3.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.0
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.15
marshmallow==3.23.1
matplotlib==3.9.2
msgpack==1.1.0
openai==1.55.3
orjson==3.10.12
protobuf==4.25.3
pyasn1-modules==0.2.8
pyls-spyder==0.4.0
pypdf==5.1.0
PyQt5==5.15.10
PyQtWebEngine==5.15.6
setuptools==75.1.0
streamlit==1.40.2
striprtf==0.0.26
tiktoken==0.8.0
toolhouse==1.3.0
typing-inspect==0.9.0
wheel==0.44.0